+++ _sample_suites/interrupted_session/test_1.py
@@ -1,0 +1,10 @@
+import gossip
+
+
+@gossip.register('slash.session_start')
+def session_start():
+    raise KeyboardInterrupt()
+
+
+def test_1():
+    pass